TP/TP2/requirements.txt
matplotlib==3.10.6
numpy==2.3.3
Pillow==11.3.0
skimage==0.0
